# Windows: improve checkout performance with large numbers of files